.history/requirements_20241203161641.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
pydantic==2.6.1
python-dotenv==1.0.0
mysqlclient==2.2.1
passlib==1.7.4
python-jose==3.3.0
pydantic-settings==2.1.0